Updated keyboard shortcuts to work with new webviews. Added support for changing keyboard shortcuts in the code. I still have to add a settings page for them though. Moved a bunch of code to separate files for ObservableObjects
