+++ iPad browser/Views/Main/ContentView.swift
@@ -13,7 +13,6 @@ import AppKit
 #endif
 import WebKit
 import Combine
-import FaviconFinder
 import SDWebImage
 import SDWebImageSwiftUI
 import SwiftData
@@ -37,15 +36,15 @@ struct ContentView: View {
     @StateObject var manager =  WebsiteManager()
     
     @AppStorage("currentSpace") var currentSpace = "Untitled"
-
-        let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
-        let rotationTimer = Timer.publish(every: 1.5, on: .main, in: .common).autoconnect()
-
-        @AppStorage("startColorHex") var startHex = "8A3CEF"
-        @AppStorage("endColorHex") var endHex = "84F5FE"
-        @AppStorage("textColorHex") var textHex = "ffffff"
-
-        
+    
+    let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
+    let rotationTimer = Timer.publish(every: 1.5, on: .main, in: .common).autoconnect()
+    
+    @AppStorage("startColorHex") var startHex = "8A3CEF"
+    @AppStorage("endColorHex") var endHex = "84F5FE"
+    @AppStorage("textColorHex") var textHex = "ffffff"
+    
+    
     @State var boostEditor = false
     
     @AppStorage("hideSidebar") var hideSidebar = false
@@ -63,570 +62,42 @@ struct ContentView: View {
     
     var body: some View {
         ZStack {
-#if !os(macOS)
-        if UIDevice.current.userInterfaceIdiom != .phone {
-            GeometryReader { geo in
-                if spaces.count > 0 {
-                    ZStack {
+            if UIDevice.current.userInterfaceIdiom != .phone {
+                GeometryReader { geo in
+                    if spaces.count > 0 {
                         ZStack {
+                            ZStack {
 #if !os(visionOS)
-                            if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
-                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                            }
-                            else {
-                                LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                            }
-#endif
-                            
-                            if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                Color.black.opacity(0.5)
-                            }
-                            
-                            
-                            HStack(spacing: 0) {
-                                //if UIDevice.current.userInterfaceIdiom != .phone {
-                                if true {
-                                    if settings.sidebarLeft {
-                                        if settings.showBorder {
-                                            ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
-                                                .disabled(true)
-                                        }
-                                        
-                                        if settings.swipingSpaces {
-                                            PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
-                                        }
-                                        else {
-                                            VStack {
-                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                    .padding([.top, .horizontal], 5)
-                                                
-                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                
-                                                HStack {
-                                                    Button {
-                                                        variables.showSettings.toggle()
-                                                    } label: {
-                                                        ZStack {
-                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                            
-                                                            Image(systemName: "gearshape")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.settingsButtonHover = true
-                                                                }
-                                                                else {
-                                                                    variables.settingsButtonHover = false
-                                                                }
-                                                            })
-                                                    }
-                                                    .sheet(isPresented: $variables.showSettings) {
-                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                .presentationSizing(.form)
-                                                        } else {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                        }
-                                                    }
-                                                    Spacer()
-                                                    
-                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                    
-                                                    Button(action: {
-                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                    }, label: {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            Color(.white)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
-#endif
-                                                            
-                                                            Image(systemName: "plus")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                }
-                                                                else {
-                                                                    variables.hoverSpace = ""
-                                                                }
-                                                            })
-                                                    })
-                                                }
-                                            }.frame(width: 300)
-                                        }
-                                    }
+                                if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
+                                    LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
                                 }
-                                HStack {
-                                    GeometryReader { webGeo in
-                                        ZStack {
-                                            Color.white
-                                                .opacity(0.4)
-                                                .cornerRadius(10)
-                                            
-                                            
-                                            //MARK: - WebView
-                                            CurrentWebView(webGeo: webGeo)
-                                                .overlay(content: {
-                                                    loadingIndicators(for: manager.selectedWebView?.webView.isLoading ?? false)
-                                                })
-                                            
-                                            if !settings.swipeNavigationDisabled {
-                                                if manager.selectedWebView != nil {
-                                                    HStack(alignment: .center, spacing: 0) {
-                                                        navigationButton(imageName: "arrow.left", action: {
-                                                            manager.selectedWebView?.webView.goBack()
-                                                        })
-                                                            .padding(.trailing, 30)
-                                                        
-                                                        Spacer()
-                                                            .frame(width: webGeo.size.width)
-                                                        
-                                                        navigationButton(imageName: "arrow.right", action: {
-                                                            manager.selectedWebView?.webView.goForward()
-                                                        })
-                                                            .padding(.leading, 30)
-                                                        
-                                                    }
-                                                    .frame(width: webGeo.size.width)
-                                                    .offset(x: variables.navigationOffset)
-                                                }
-                                            }
-                                            
-                                            if variables.auraTab == "dashboard" && variables.selectedTabLocation == "" {
-                                                Dashboard(startHexSpace: spaces[selectedSpaceIndex].startHex, endHexSpace: spaces[selectedSpaceIndex].endHex)
-                                                    .cornerRadius(10)
-                                                    .clipped()
-                                            }
-                                            
-                                            Spacer()
-                                                .frame(width: 20)
-                                            
-                                            HStack {
-                                                Button {
-                                                    Task {
-                                                        await hideSidebar.toggle()
-                                                    }
-                                                    
-                                                    withAnimation {
-                                                        if !hideSidebar {
-                                                            if variables.selectedTabLocation == "tabs" {
-                                                                Task {
-                                                                    await variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "pinnedTabs" {
-                                                                Task {
-                                                                    await variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "favoriteTabs" {
-                                                                Task {
-                                                                    await variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                        } else {
-                                                            if variables.selectedTabLocation == "tabs" {
-                                                                Task {
-                                                                    await variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "pinnedTabs" {
-                                                                Task {
-                                                                    await variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "favoriteTabs" {
-                                                                Task {
-                                                                    await variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("s", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.goBack()
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.goBack()
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.goBack()
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("[", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.goForward()
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.goForward()
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.goForward()
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("]", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    variables.reloadRotation += 360
-                                                    
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.reload()
-                                                        variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.navigationState.selectedWebView = variables.navigationState.selectedWebView
-                                                        //navigationState.currentURL = navigationState.currentURL
-                                                        
-                                                        if let unwrappedURL = variables.navigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.reload()
-                                                        variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.pinnedNavigationState.selectedWebView = variables.pinnedNavigationState.selectedWebView
-                                                        
-                                                        if let unwrappedURL = variables.pinnedNavigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.reload()
-                                                        variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.favoritesNavigationState.selectedWebView = variables.favoritesNavigationState.selectedWebView
-                                                        
-                                                        if let unwrappedURL = variables.favoritesNavigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("r", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    if (variables.navigationState.selectedWebView == nil) && (variables.pinnedNavigationState.selectedWebView == nil) && (variables.favoritesNavigationState.selectedWebView == nil) {
-                                                        variables.tabBarShown.toggle()
-                                                        variables.commandBarShown = false
-                                                        print("Showing Tab Bar")
-                                                    }
-                                                    else {
-                                                        if variables.selectedTabLocation == "pinnedTabs" {
-                                                            variables.searchInSidebar = unformatURL(url: variables.pinnedNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }
-                                                        else if variables.selectedTabLocation == "favoriteTabs" {
-                                                            variables.searchInSidebar = unformatURL(url: variables.favoritesNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }else {
-                                                            variables.searchInSidebar = unformatURL(url: variables.navigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }
-                                                        variables.commandBarShown.toggle()
-                                                        variables.tabBarShown = false
-                                                        print("Showing Command Bar")
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("l", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "favoriteTabs" {
-                                                        if let index = variables.favoritesNavigationState.webViews.firstIndex(of: variables.favoritesNavigationState.selectedWebView ?? WKWebView()) {
-                                                            favoriteRemoveTab(at: index)
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        if let index = variables.pinnedNavigationState.webViews.firstIndex(of: variables.pinnedNavigationState.selectedWebView ?? WKWebView()) {
-                                                            pinnedRemoveTab(at: index)
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "tabs" {
-                                                        if let index = variables.navigationState.webViews.firstIndex(of: variables.navigationState.selectedWebView ?? WKWebView()) {
-                                                            removeTab(at: index)
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("w", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    variables.tabBarShown.toggle()
-                                                    variables.commandBarShown = false
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("t", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                            }
-                                        }
-                                        .gesture(
-                                            DragGesture()
-                                                .onChanged { value in
-                                                    if !settings.swipeNavigationDisabled {
-                                                        let startLocation = value.startLocation.x
-                                                        let width = webGeo.size.width
-                                                        
-                                                        if startLocation < 100 || startLocation > (width - 100) {
-                                                            let newOffset = value.translation.width
-                                                            if abs(newOffset) <= 150 {
-                                                                variables.navigationOffset = newOffset
-                                                            } else {
-                                                                variables.navigationOffset = newOffset > 0 ? 150 : -150
-                                                            }
-                                                            if abs(newOffset) > 100 {
-                                                                if !variables.arrowImpactOnce {
-                                                                    heavyHaptics()
-                                                                    
-                                                                    variables.arrowImpactOnce = true
-                                                                }
-                                                                
-                                                                withAnimation(.linear(duration: 0.3)) {
-                                                                    variables.navigationArrowColor = true
-                                                                }
-                                                            } else {
-                                                                if variables.arrowImpactOnce {
-                                                                    heavyHaptics()
-                                                                    
-                                                                    variables.arrowImpactOnce = false
-                                                                }
-                                                                withAnimation(.linear(duration: 0.3)) {
-                                                                    variables.navigationArrowColor = false
-                                                                }
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                                .onEnded { value in
-                                                    if !settings.swipeNavigationDisabled {
-                                                        variables.arrowImpactOnce = false
-                                                        let startLocation = value.startLocation.x
-                                                        let width = webGeo.size.width
-                                                        
-                                                        if startLocation < 150 || startLocation > (width - 150) {
-                                                            if variables.navigationOffset >= 100 {
-                                                                goBack()
-                                                            } else if variables.navigationOffset < -100 {
-                                                                goForward()
-                                                            }
-                                                            
-                                                            withAnimation(.linear(duration: 0.25)) {
-                                                                variables.navigationOffset = 0
-                                                                variables.navigationArrowColor = false
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                        )
-                                        .onChange(of: webGeo.size) {
-                                            variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                            
-                                            variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                            
-                                            variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                        }
-                                    }
-                                    .cornerRadius(10)
-                                    .clipped()
-                                    .padding(settings.sidebarLeft ? .trailing: .leading, settings.showBorder ? 12: 0)
-                                    
-                                    if variables.delayedBrowseForMe {
-                                        BrowseForMe(searchText: variables.browseForMeSearch, searchResponse: "", closeSheet: $variables.isBrowseForMe)
-                                            .frame(width: variables.isBrowseForMe ? 400: 0)
-                                            .cornerRadius(10)
-                                            .clipped()
-                                    }
-                                }.onChange(of: variables.isBrowseForMe, {
-                                    DispatchQueue.main.asyncAfter(deadline: .now() + 1, execute: {
-                                        withAnimation(.linear, {
-                                            variables.delayedBrowseForMe = variables.isBrowseForMe
-                                        })
-                                    })
-                                })
-                                //if UIDevice.current.userInterfaceIdiom != .phone {
-                                if true {
-                                    if !settings.sidebarLeft {
-                                        if settings.showBorder {
-                                            ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
-                                                .disabled(true)
-                                        }
-                                        
-                                        if settings.swipingSpaces {
-                                            PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
-                                        }
-                                        else {
-                                            VStack {
-                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                    .padding([.top, .horizontal], 5)
-                                                
-                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                
-                                                HStack {
-                                                    Button {
-                                                        variables.showSettings.toggle()
-                                                    } label: {
-                                                        ZStack {
-                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                            
-                                                            Image(systemName: "gearshape")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.settingsButtonHover = true
-                                                                }
-                                                                else {
-                                                                    variables.settingsButtonHover = false
-                                                                }
-                                                            })
-                                                    }
-                                                    .sheet(isPresented: $variables.showSettings) {
-                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                .presentationSizing(.form)
-                                                        } else {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                        }
-                                                    }
-                                                    Spacer()
-                                                    
-                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                    
-                                                    Button(action: {
-                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                    }, label: {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            Color(.white)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
-#endif
-                                                            
-                                                            Image(systemName: "plus")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                }
-                                                                else {
-                                                                    variables.hoverSpace = ""
-                                                                }
-                                                            })
-                                                    })
-                                                }
-                                            }.frame(width: 300)
-                                        }
-                                    }
+                                else {
+                                    LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
                                 }
-                            }
-                            .padding(.trailing, settings.showBorder ? 10: 0)
-                            .padding(.vertical, settings.showBorder ? 25: 0)
-                            .onAppear {
-                                if let savedStartColor = getColor(forKey: "startColorHex") {
-                                    variables.startColor = savedStartColor
-                                }
-                                if let savedEndColor = getColor(forKey: "endColorHex") {
-                                    variables.endColor = savedEndColor
-                                }
-                                if let savedTextColor = getColor(forKey: "textColorHex") {
-                                    variables.textColor = savedTextColor
+#endif
+                                
+                                if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
+                                    Color.black.opacity(0.5)
                                 }
                                 
-                                variables.spaceIcons = UserDefaults.standard.dictionary(forKey: "spaceIcons") as? [String: String]
-                            }
-                            
-                            if variables.tabBarShown || variables.commandBarShown || variables.tapSidebarShown {
-                                Button(action: {
-                                    variables.tabBarShown = false
-                                    variables.commandBarShown = false
-                                    variables.tapSidebarShown = false
-                                }, label: {
-                                    Color.white.opacity(0.0001)
-                                }).buttonStyle(.plain)
-#if !os(macOS)
-                                    .hoverEffectDisabled(true)
-                                #endif
-                            }
-                            
-                            if hideSidebar {
-                                HStack {
-                                    if !settings.sidebarLeft {
-                                        Spacer()
-                                    }
-                                    
-                                    ZStack {
-                                        Color.white.opacity(0.00001)
-                                            .frame(width: 35)
-                                            .onTapGesture {
-                                                variables.tapSidebarShown = true
-                                                
+                                
+                                HStack(spacing: 0) {
+                                    //if UIDevice.current.userInterfaceIdiom != .phone {
+                                    if true {
+                                        if settings.sidebarLeft {
+                                            if settings.showBorder {
+                                                ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
+                                                    .disabled(true)
                                             }
-                                        
-                                        HStack {
-                                            //                                        PagedSidebar(selectedTabLocation: $selectedTabLocation, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
                                             
-                                            //if UIDevice.current.userInterfaceIdiom != .phone {
-                                            if true {
+                                            if settings.swipingSpaces {
+                                                PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
+                                            }
+                                            else {
                                                 VStack {
                                                     ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
                                                         .padding([.top, .horizontal], 5)
                                                     
-                                                    //Sidebar(selectedTabLocation: $selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
                                                     Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
                                                     
                                                     HStack {
@@ -700,1371 +171,854 @@ struct ContentView: View {
                                                                 })
                                                         })
                                                     }
+                                                }.frame(width: 300)
+                                            }
+                                        }
+                                    }
+                                    HStack {
+                                        GeometryReader { webGeo in
+                                            ZStack {
+                                                Color.white
+                                                    .opacity(0.4)
+                                                    .cornerRadius(10)
+                                                
+                                                
+                                                //MARK: - WebView
+                                                CurrentWebView(webGeo: webGeo)
+                                                    .overlay(content: {
+                                                        loadingIndicators(for: manager.selectedWebView?.webView.isLoading ?? false)
+                                                    })
+                                                
+                                                if !settings.swipeNavigationDisabled {
+                                                    if manager.selectedWebView != nil {
+                                                        HStack(alignment: .center, spacing: 0) {
+                                                            navigationButton(imageName: "arrow.left", action: {
+                                                                manager.selectedWebView?.webView.goBack()
+                                                            })
+                                                            .padding(.trailing, 30)
+                                                            
+                                                            Spacer()
+                                                                .frame(width: webGeo.size.width)
+                                                            
+                                                            navigationButton(imageName: "arrow.right", action: {
+                                                                manager.selectedWebView?.webView.goForward()
+                                                            })
+                                                            .padding(.leading, 30)
+                                                            
+                                                        }
+                                                        .frame(width: webGeo.size.width)
+                                                        .offset(x: variables.navigationOffset)
+                                                    }
                                                 }
-                                                .padding(15)
-                                                .frame(width: 300)
-                                                .background(content: {
-#if !os(visionOS)
-                                                    if settings.sidebarLeft {
-                                                        LinearGradient(colors: [variables.startColor, Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            .opacity(1.0)
-                                                        if selectedSpaceIndex < spaces.count {
-                                                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
+                                                
+                                                if variables.auraTab == "dashboard" && variables.selectedTabLocation == "" {
+                                                    Dashboard(startHexSpace: spaces[selectedSpaceIndex].startHex, endHexSpace: spaces[selectedSpaceIndex].endHex)
+                                                        .cornerRadius(10)
+                                                        .clipped()
+                                                }
+                                                
+                                                Spacer()
+                                                    .frame(width: 20)
+                                                
+                                                HStack {
+                                                    Button {
+                                                        Task {
+                                                            await hideSidebar.toggle()
                                                         }
-                                                    } else {
-                                                        LinearGradient(colors: [Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex)), variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            .opacity(1.0)
-                                                        if selectedSpaceIndex < spaces.count {
-                                                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                                                LinearGradient(colors: [Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex)), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-                                                        }
+                                                    } label: {
+                                                        
                                                     }
-#endif
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.toggleSidebar))
+                                                    .buttonStyle(.plain)
                                                     
-                                                    if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                                        Color.black.opacity(0.5)
+                                                    Button {
+                                                        manager.selectedWebView?.webView.goBack()
+                                                    } label: {
+                                                        
                                                     }
-                                                })
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.goBack))
+                                                    .buttonStyle(.plain)
+                                                    
+                                                    Button {
+                                                        manager.selectedWebView?.webView.goForward()
+                                                    } label: {
+                                                        
+                                                    }
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.goForward))
+                                                    .buttonStyle(.plain)
+                                                    
+                                                    
+                                                    Button {
+                                                        variables.reloadRotation += 360
+                                                        
+                                                        if variables.selectedTabLocation == "tabs" {
+                                                            variables.navigationState.selectedWebView?.reload()
+                                                            variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
+                                                            
+                                                            variables.navigationState.selectedWebView = variables.navigationState.selectedWebView
+                                                            //navigationState.currentURL = navigationState.currentURL
+                                                            
+                                                            if let unwrappedURL = variables.navigationState.currentURL {
+                                                                variables.searchInSidebar = unwrappedURL.absoluteString
+                                                            }
+                                                        }
+                                                        else if variables.selectedTabLocation == "pinnedTabs" {
+                                                            variables.pinnedNavigationState.selectedWebView?.reload()
+                                                            variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
+                                                            
+                                                            variables.pinnedNavigationState.selectedWebView = variables.pinnedNavigationState.selectedWebView
+                                                            
+                                                            if let unwrappedURL = variables.pinnedNavigationState.currentURL {
+                                                                variables.searchInSidebar = unwrappedURL.absoluteString
+                                                            }
+                                                        }
+                                                        else if variables.selectedTabLocation == "favoriteTabs" {
+                                                            variables.favoritesNavigationState.selectedWebView?.reload()
+                                                            variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
+                                                            
+                                                            variables.favoritesNavigationState.selectedWebView = variables.favoritesNavigationState.selectedWebView
+                                                            
+                                                            if let unwrappedURL = variables.favoritesNavigationState.currentURL {
+                                                                variables.searchInSidebar = unwrappedURL.absoluteString
+                                                            }
+                                                        }
+                                                    } label: {
+                                                        
+                                                    }
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.reload))
+                                                    .buttonStyle(.plain)
+                                                    
+                                                    
+                                                    Button {
+                                                        if (variables.navigationState.selectedWebView == nil) && (variables.pinnedNavigationState.selectedWebView == nil) && (variables.favoritesNavigationState.selectedWebView == nil) {
+                                                            variables.tabBarShown.toggle()
+                                                            variables.commandBarShown = false
+                                                            print("Showing Tab Bar")
+                                                        }
+                                                        else {
+                                                            if variables.selectedTabLocation == "pinnedTabs" {
+                                                                variables.searchInSidebar = unformatURL(url: variables.pinnedNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
+                                                            }
+                                                            else if variables.selectedTabLocation == "favoriteTabs" {
+                                                                variables.searchInSidebar = unformatURL(url: variables.favoritesNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
+                                                            }else {
+                                                                variables.searchInSidebar = unformatURL(url: variables.navigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
+                                                            }
+                                                            variables.commandBarShown.toggle()
+                                                            variables.tabBarShown = false
+                                                            print("Showing Command Bar")
+                                                        }
+                                                    } label: {
+                                                        
+                                                    }
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.commandBar))
+                                                    .buttonStyle(.plain)
+                                                    
+                                                    
+                                                    Button {
+                                                        if variables.selectedTabLocation == "favoriteTabs" {
+                                                            if let index = variables.favoritesNavigationState.webViews.firstIndex(of: variables.favoritesNavigationState.selectedWebView ?? WKWebView()) {
+                                                                favoriteRemoveTab(at: index)
+                                                            }
+                                                        }
+                                                        else if variables.selectedTabLocation == "pinnedTabs" {
+                                                            if let index = variables.pinnedNavigationState.webViews.firstIndex(of: variables.pinnedNavigationState.selectedWebView ?? WKWebView()) {
+                                                                pinnedRemoveTab(at: index)
+                                                            }
+                                                        }
+                                                        else if variables.selectedTabLocation == "tabs" {
+                                                            if let index = variables.navigationState.webViews.firstIndex(of: variables.navigationState.selectedWebView ?? WKWebView()) {
+                                                                removeTab(at: index)
+                                                            }
+                                                        }
+                                                    } label: {
+                                                        
+                                                    }.keyboardShortcut("w", modifiers: .command)
+                                                        .buttonStyle(.plain)
+                                                    
+                                                    
+                                                    Button {
+                                                        variables.tabBarShown.toggle()
+                                                        variables.commandBarShown = false
+                                                    } label: {
+                                                        
+                                                    }
+                                                    .keyboardShortcut(variables.shortcuts.parseShortcut(shortcut: variables.shortcuts.newTab))
+                                                    .buttonStyle(.plain)
+                                                    
+                                                    
+                                                }
+                                            }
+                                            .gesture(
+                                                DragGesture()
+                                                    .onChanged { value in
+                                                        if !settings.swipeNavigationDisabled {
+                                                            let startLocation = value.startLocation.x
+                                                            let width = webGeo.size.width
+                                                            
+                                                            if startLocation < 100 || startLocation > (width - 100) {
+                                                                let newOffset = value.translation.width
+                                                                if abs(newOffset) <= 150 {
+                                                                    variables.navigationOffset = newOffset
+                                                                } else {
+                                                                    variables.navigationOffset = newOffset > 0 ? 150 : -150
+                                                                }
+                                                                if abs(newOffset) > 100 {
+                                                                    if !variables.arrowImpactOnce {
+                                                                        heavyHaptics()
+                                                                        
+                                                                        variables.arrowImpactOnce = true
+                                                                    }
+                                                                    
+                                                                    withAnimation(.linear(duration: 0.3)) {
+                                                                        variables.navigationArrowColor = true
+                                                                    }
+                                                                } else {
+                                                                    if variables.arrowImpactOnce {
+                                                                        heavyHaptics()
+                                                                        
+                                                                        variables.arrowImpactOnce = false
+                                                                    }
+                                                                    withAnimation(.linear(duration: 0.3)) {
+                                                                        variables.navigationArrowColor = false
+                                                                    }
+                                                                }
+                                                            }
+                                                        }
+                                                    }
+                                                    .onEnded { value in
+                                                        if !settings.swipeNavigationDisabled {
+                                                            variables.arrowImpactOnce = false
+                                                            let startLocation = value.startLocation.x
+                                                            let width = webGeo.size.width
+                                                            
+                                                            if startLocation < 150 || startLocation > (width - 150) {
+                                                                if variables.navigationOffset >= 100 {
+                                                                    goBack()
+                                                                } else if variables.navigationOffset < -100 {
+                                                                    goForward()
+                                                                }
+                                                                
+                                                                withAnimation(.linear(duration: 0.25)) {
+                                                                    variables.navigationOffset = 0
+                                                                    variables.navigationArrowColor = false
+                                                                }
+                                                            }
+                                                        }
+                                                    }
+                                            )
+                                            .onChange(of: webGeo.size) {
+                                                variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
+                                                
+                                                variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
+                                                
+                                                variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
+                                            }
+                                        }
+                                        .cornerRadius(10)
+                                        .clipped()
+                                        .padding(settings.sidebarLeft ? .trailing: .leading, settings.showBorder ? 12: 0)
+                                        
+                                        if variables.delayedBrowseForMe {
+                                            BrowseForMe(searchText: variables.browseForMeSearch, searchResponse: "", closeSheet: $variables.isBrowseForMe)
+                                                .frame(width: variables.isBrowseForMe ? 400: 0)
                                                 .cornerRadius(10)
-                                                .shadow(color: .black.opacity(0.25), radius: 10, x: 0, y: 0)
+                                                .clipped()
+                                        }
+                                    }.onChange(of: variables.isBrowseForMe, {
+                                        DispatchQueue.main.asyncAfter(deadline: .now() + 1, execute: {
+                                            withAnimation(.linear, {
+                                                variables.delayedBrowseForMe = variables.isBrowseForMe
+                                            })
+                                        })
+                                    })
+                                    //if UIDevice.current.userInterfaceIdiom != .phone {
+                                    if true {
+                                        if !settings.sidebarLeft {
+                                            if settings.showBorder {
+                                                ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
+                                                    .disabled(true)
+                                            }
+                                            
+                                            if settings.swipingSpaces {
+                                                PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
                                             }
                                             else {
-                                                Color.clear
-                                                    .sheet(isPresented: $variables.tapSidebarShown) {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
-                                                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-                                                            else {
-                                                                LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-#endif
-                                                            
-                                                            if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                                                Color.black.opacity(0.5)
-                                                            }
-                                                            
-                                                            VStack {
-                                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                                    .padding([.top, .horizontal], 5)
+                                                VStack {
+                                                    ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
+                                                        .padding([.top, .horizontal], 5)
+                                                    
+                                                    Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
+                                                    
+                                                    HStack {
+                                                        Button {
+                                                            variables.showSettings.toggle()
+                                                        } label: {
+                                                            ZStack {
+                                                                HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
                                                                 
-                                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
+                                                                Image(systemName: "gearshape")
+                                                                    .resizable()
+                                                                    .scaledToFit()
+                                                                    .frame(width: 20, height: 20)
+                                                                    .foregroundStyle(variables.textColor)
+                                                                    .opacity(variables.settingsButtonHover ? 1.0: 0.5)
                                                                 
-                                                                HStack {
-                                                                    Button {
-                                                                        variables.showSettings.toggle()
-                                                                    } label: {
-                                                                        ZStack {
-                                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                                            
-                                                                            Image(systemName: "gearshape")
-                                                                                .resizable()
-                                                                                .scaledToFit()
-                                                                                .frame(width: 20, height: 20)
-                                                                                .foregroundStyle(variables.textColor)
-                                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                                            
-                                                                        }.frame(width: 40, height: 40).cornerRadius(7)
+                                                            }.frame(width: 40, height: 40).cornerRadius(7)
 #if !os(visionOS) && !os(macOS)
-                                                                            .hoverEffect(.lift)
-                                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+                                                                .hoverEffect(.lift)
+                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
 #endif
-                                                                            .onHover(perform: { hovering in
-                                                                                if hovering {
-                                                                                    variables.settingsButtonHover = true
-                                                                                }
-                                                                                else {
-                                                                                    variables.settingsButtonHover = false
-                                                                                }
-                                                                            })
+                                                                .onHover(perform: { hovering in
+                                                                    if hovering {
+                                                                        variables.settingsButtonHover = true
                                                                     }
-                                                                    .sheet(isPresented: $variables.showSettings) {
-                                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                                .presentationSizing(.form)
-                                                                        } else {
-                                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                                    else {
+                                                                        variables.settingsButtonHover = false
+                                                                    }
+                                                                })
+                                                        }
+                                                        .sheet(isPresented: $variables.showSettings) {
+                                                            if #available(iOS 18.0, visionOS 2.0, *) {
+                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                                    .presentationSizing(.form)
+                                                            } else {
+                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                            }
+                                                        }
+                                                        Spacer()
+                                                        
+                                                        SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
+                                                        
+                                                        Button(action: {
+                                                            modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
+                                                        }, label: {
+                                                            ZStack {
+#if !os(visionOS)
+                                                                Color(.white)
+                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
+#endif
+                                                                
+                                                                Image(systemName: "plus")
+                                                                    .resizable()
+                                                                    .scaledToFit()
+                                                                    .frame(width: 20, height: 20)
+                                                                    .foregroundStyle(variables.textColor)
+                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
+                                                                
+                                                            }.frame(width: 40, height: 40).cornerRadius(7)
+#if !os(visionOS) && !os(macOS)
+                                                                .hoverEffect(.lift)
+                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+#endif
+                                                                .onHover(perform: { hovering in
+                                                                    if hovering {
+                                                                        variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
+                                                                    }
+                                                                    else {
+                                                                        variables.hoverSpace = ""
+                                                                    }
+                                                                })
+                                                        })
+                                                    }
+                                                }.frame(width: 300)
+                                            }
+                                        }
+                                    }
+                                }
+                                .padding(.trailing, settings.showBorder ? 10: 0)
+                                .padding(.vertical, settings.showBorder ? 25: 0)
+                                .onAppear {
+                                    if let savedStartColor = getColor(forKey: "startColorHex") {
+                                        variables.startColor = savedStartColor
+                                    }
+                                    if let savedEndColor = getColor(forKey: "endColorHex") {
+                                        variables.endColor = savedEndColor
+                                    }
+                                    if let savedTextColor = getColor(forKey: "textColorHex") {
+                                        variables.textColor = savedTextColor
+                                    }
+                                    
+                                    variables.spaceIcons = UserDefaults.standard.dictionary(forKey: "spaceIcons") as? [String: String]
+                                }
+                                
+                                if variables.tabBarShown || variables.commandBarShown || variables.tapSidebarShown {
+                                    Button(action: {
+                                        variables.tabBarShown = false
+                                        variables.commandBarShown = false
+                                        variables.tapSidebarShown = false
+                                    }, label: {
+                                        Color.white.opacity(0.0001)
+                                    }).buttonStyle(.plain)
+#if !os(macOS)
+                                        .hoverEffectDisabled(true)
+#endif
+                                }
+                                
+                                if hideSidebar {
+                                    HStack {
+                                        if !settings.sidebarLeft {
+                                            Spacer()
+                                        }
+                                        
+                                        ZStack {
+                                            Color.white.opacity(0.00001)
+                                                .frame(width: 35)
+                                                .onTapGesture {
+                                                    variables.tapSidebarShown = true
+                                                    
+                                                }
+                                            
+                                            HStack {
+                                                //                                        PagedSidebar(selectedTabLocation: $selectedTabLocation, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
+                                                
+                                                //if UIDevice.current.userInterfaceIdiom != .phone {
+                                                if true {
+                                                    VStack {
+                                                        ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
+                                                            .padding([.top, .horizontal], 5)
+                                                        
+                                                        //Sidebar(selectedTabLocation: $selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
+                                                        Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
+                                                        
+                                                        HStack {
+                                                            Button {
+                                                                variables.showSettings.toggle()
+                                                            } label: {
+                                                                ZStack {
+                                                                    HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
+                                                                    
+                                                                    Image(systemName: "gearshape")
+                                                                        .resizable()
+                                                                        .scaledToFit()
+                                                                        .frame(width: 20, height: 20)
+                                                                        .foregroundStyle(variables.textColor)
+                                                                        .opacity(variables.settingsButtonHover ? 1.0: 0.5)
+                                                                    
+                                                                }.frame(width: 40, height: 40).cornerRadius(7)
+#if !os(visionOS) && !os(macOS)
+                                                                    .hoverEffect(.lift)
+                                                                    .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+#endif
+                                                                    .onHover(perform: { hovering in
+                                                                        if hovering {
+                                                                            variables.settingsButtonHover = true
                                                                         }
-                                                                    }
-                                                                    Spacer()
-                                                                    
-                                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                                    
-                                                                    Button(action: {
-                                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                                    }, label: {
-                                                                        ZStack {
-#if !os(visionOS)
-                                                                            Color(.white)
-                                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
-#endif
-                                                                            
-                                                                            Image(systemName: "plus")
-                                                                                .resizable()
-                                                                                .scaledToFit()
-                                                                                .frame(width: 20, height: 20)
-                                                                                .foregroundStyle(variables.textColor)
-                                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                                            
-                                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                                            .hoverEffect(.lift)
-                                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                                            .onHover(perform: { hovering in
-                                                                                if hovering {
-                                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                                }
-                                                                                else {
-                                                                                    variables.hoverSpace = ""
-                                                                                }
-                                                                            })
+                                                                        else {
+                                                                            variables.settingsButtonHover = false
+                                                                        }
                                                                     })
+                                                            }
+                                                            .sheet(isPresented: $variables.showSettings) {
+                                                                if #available(iOS 18.0, visionOS 2.0, *) {
+                                                                    NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                                        .presentationSizing(.form)
+                                                                } else {
+                                                                    NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
                                                                 }
                                                             }
-                                                            .padding(15)
-                                                            .frame(width: 300)
-                                                            .cornerRadius(10)
+                                                            Spacer()
+                                                            
+                                                            SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
+                                                            
+                                                            Button(action: {
+                                                                modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
+                                                            }, label: {
+                                                                ZStack {
+#if !os(visionOS)
+                                                                    Color(.white)
+                                                                        .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
+#endif
+                                                                    
+                                                                    Image(systemName: "plus")
+                                                                        .resizable()
+                                                                        .scaledToFit()
+                                                                        .frame(width: 20, height: 20)
+                                                                        .foregroundStyle(variables.textColor)
+                                                                        .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
+                                                                    
+                                                                }.frame(width: 40, height: 40).cornerRadius(7)
+#if !os(visionOS) && !os(macOS)
+                                                                    .hoverEffect(.lift)
+                                                                    .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+#endif
+                                                                    .onHover(perform: { hovering in
+                                                                        if hovering {
+                                                                            variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
+                                                                        }
+                                                                        else {
+                                                                            variables.hoverSpace = ""
+                                                                        }
+                                                                    })
+                                                            })
                                                         }
                                                     }
-                                            }
-                                            
-                                            Spacer()
-                                        }.padding(40)
-                                            .padding(.leading, 30)
-                                            .frame(width: variables.hoveringSidebar || variables.tapSidebarShown ? 350: 0)
-                                            .offset(x: variables.hoveringSidebar || variables.tapSidebarShown ? 0: settings.sidebarLeft ? -350: 300)
-                                            .clipped()
-                                        
-                                    }.onHover(perform: { hovering in
-                                        if hovering {
-                                            variables.hoveringSidebar = true
-                                        }
-                                        else {
-                                            variables.hoveringSidebar = false
-                                        }
-                                    })
-                                    
-                                    if settings.sidebarLeft {
-                                        Spacer()
-                                    }
-                                }.animation(.default)
-                            }
-                        }
-                        .onOpenURL { url in
-                            if url.absoluteString.starts(with: "aura://") {
-                                //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: "https\(url.absoluteString.dropFirst(4))")!))
-                            }
-                            else {
-                                //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: url.absoluteString)!))
-                                
-                                spaces[selectedSpaceIndex].tabUrls.append(url.absoluteString)
-                            }
-                            print("Url:")
-                            print(url)
-                        }
-                        //MARK: - Tabbar
-                        if variables.tabBarShown {
-                            CommandBar(commandBarText: $variables.newTabSearch, searchSubmitted: $variables.commandBarSearchSubmitted, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
-                                .onChange(of: variables.commandBarSearchSubmitted) { thing in
-                                    
-                                    print("Search submitted")
-                                    
-                                    variables.browseForMeSearch = variables.newTabSearch
-                                    
-                                    if !variables.newTabSearch.starts(with: "aura://") {
-                                        variables.auraTab = ""
-                                        //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: formatURL(from: variables.newTabSearch))!))
-                                        
-                                        print(variables.newTabSearch)
-                                        
-                                        var temporaryUrls = spaces[selectedSpaceIndex].tabUrls
-                                        
-                                        print("temporaryUrls:")
-                                        print(temporaryUrls)
-                                        
-                                        var formattedUrl = formatURL(from: variables.newTabSearch)
-                                        
-                                        print("formattedUrl:")
-                                        print(formattedUrl)
-                                        
-                                        temporaryUrls.append(formattedUrl)
-                                        
-                                        print("temporaryUrls - changed:")
-                                        print(temporaryUrls)
-                                        spaces[selectedSpaceIndex].tabUrls = temporaryUrls
-                                        
-                                        print("tabUrls:")
-                                        print(spaces[selectedSpaceIndex].tabUrls)
-                                        
-                                        do {
-                                            try modelContext.save()
-                                        } catch {
-                                            print(error.localizedDescription)
-                                        }
-                                        print(spaces[selectedSpaceIndex].tabUrls)
-                                    }
-                                    else {
-                                        if variables.newTabSearch.contains("dashboard") {
-                                            variables.navigationState.selectedWebView = nil
-                                            variables.pinnedNavigationState.selectedWebView = nil
-                                            variables.favoritesNavigationState.selectedWebView = nil
-                                            
-                                            variables.auraTab = "dashboard"
-                                            variables.selectedTabLocation = ""
-                                        }
-                                        if variables.newTabSearch.contains("settings") {
-                                            variables.showSettings = true
-                                        }
-                                    }
-                                    
-                                    variables.tabBarShown = false
-                                    variables.commandBarSearchSubmitted = false
-                                    variables.newTabSearch = ""
-                                    
-                                    print("Saving Tabs")
-                                    
-                                    //saveSpaceData()
-                                }
-                            //}
-                        }
-                        
-                        //MARK: - Command Bar
-                        else if variables.commandBarShown {
-                            CommandBar(commandBarText: $variables.searchInSidebar, searchSubmitted: $variables.commandBarSearchSubmitted2, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
-                                .onChange(of: variables.navigationState.currentURL, {
-                                    if let unwrappedURL = variables.navigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 1")
-                                })
-                                .onChange(of: variables.pinnedNavigationState.currentURL, {
-                                    if let unwrappedURL = variables.pinnedNavigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 2")
-                                })
-                                .onChange(of: variables.favoritesNavigationState.currentURL, {
-                                    if let unwrappedURL = variables.favoritesNavigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 3")
-                                })
-                                .onChange(of: variables.commandBarSearchSubmitted2) { thing in
-                                    
-                                    //variables.navigationState.currentURL = URL(string: formatURL(from: newTabSearch))!
-                                    //variables.navigationState.selectedWebView?.load(URLRequest(url: URL(formatURL(from: newTabSearch))!))
-                                    Task {
-                                        await variables.searchInSidebar = formatURL(from: variables.searchInSidebar)
-                                        if let url = URL(string: variables.searchInSidebar) {
-                                            // Create a URLRequest object
-                                            let request = URLRequest(url: url)
-                                            
-                                            if variables.selectedTabLocation == "tabs" {
-                                                await variables.navigationState.selectedWebView?.load(request)
-                                            }
-                                            if variables.selectedTabLocation == "pinnedTabs" {
-                                                await variables.pinnedNavigationState.selectedWebView?.load(request)
-                                            }
-                                            if variables.selectedTabLocation == "favoriteTabs" {
-                                                await variables.favoritesNavigationState.selectedWebView?.load(request)
-                                            }
-                                            
-                                            print("Updated URL String")
-                                        } else {
-                                            print("Invalid URL string")
-                                        }
-                                    }
-                                    
-                                    
-                                    variables.commandBarShown = false
-                                    variables.tabBarShown = false
-                                    variables.commandBarSearchSubmitted2 = false
-                                    variables.newTabSearch = ""
-                                }
-                        }
-                        
-                        
-                        if launchingAnimation && settings.launchAnimation {
-                            Launch_Animation()
-                        }
-                    }
-                    /*.sheet(isPresented: $isBrowseForMe, content: {
-                     VStack {
-                     if UIDevice.current.userInterfaceIdiom != .phone {
-                     BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
-                     .frame(width: geo.size.width * 0.7, height: geo.size.height * 0.9)
-                     .onDisappear() {
-                     isBrowseForMe = false
-                     newTabSearch = ""
-                     commandBarShown = false
-                     tabBarShown = false
-                     commandBarSearchSubmitted = false
-                     commandBarSearchSubmitted2 = false
-                     }
-                     }
-                     else {
-                     BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
-                     .onDisappear() {
-                     isBrowseForMe = false
-                     newTabSearch = ""
-                     commandBarShown = false
-                     tabBarShown = false
-                     commandBarSearchSubmitted = false
-                     commandBarSearchSubmitted2 = false
-                     }
-                     }
-                     }
-                     .interactiveDismissDisabled(true)
-                     })*/
-                    .onAppear() {
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 4.6, execute: {
-                            
-                        })
-                    }
-                    .onChange(of: selectedSpaceIndex, {
-                        if variables.initialLoadDone {
-                            variables.navigationState.webViews.removeAll()
-                            
-                            var reloadAuraTabs = variables.auraTab
-                            variables.auraTab = ""
-                            DispatchQueue.main.asyncAfter(deadline: .now() + 0.00001) {
-                                variables.auraTab = reloadAuraTabs
-                            }
-                            
-                            if selectedSpaceIndex < spaces.count {
-                                if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                    startHex = spaces[selectedSpaceIndex].startHex
-                                    endHex = spaces[selectedSpaceIndex].startHex
-                                    
-                                    variables.startColor = Color(hex: spaces[selectedSpaceIndex].startHex)
-                                    variables.endColor = Color(hex: spaces[selectedSpaceIndex].endHex)
-                                }
-                            }
-                            
-                            UserDefaults.standard.setValue(selectedSpaceIndex, forKey: "savedSelectedSpaceIndex")
-                        }
-                    })
-                    .onChange(of: variables.navigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving navigationState")
-                    })
-                    .onChange(of: variables.pinnedNavigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving pinnedNavigationState")
-                    })
-                    .onChange(of: variables.favoritesNavigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving favoritesNavigationState")
-                    })
-                    .onAppear() {
-                        if UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex") > spaces.count - 1 {
-                            selectedSpaceIndex = 0
-                        }
-                        else {
-                            selectedSpaceIndex = UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex")
-                        }
-                        
-                        if selectedSpaceIndex < spaces.count {
-                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                startHex = spaces[selectedSpaceIndex].startHex
-                                endHex = spaces[selectedSpaceIndex].startHex
-                                
-                                variables.startColor = Color(hex: spaces[selectedSpaceIndex].startHex)
-                                variables.endColor = Color(hex: spaces[selectedSpaceIndex].endHex)
-                            }
-                        }
-                        
-                        /*for space in spaces {
-                            if space.spaceName == currentSpace {
-                                for tab in space.tabUrls {
-                                    variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://figma.com")!))
-                                }
-                                for tab in space.pinnedUrls {
-                                    variables.pinnedNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://thebrowser.company")!))
-                                }
-                                for tab in space.favoritesUrls {
-                                    variables.favoritesNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://arc.net")!))
-                                }
-                            }
-                        }
-                        variables.navigationState.selectedWebView = nil
-                        variables.pinnedNavigationState.selectedWebView = nil
-                        variables.favoritesNavigationState.selectedWebView = nil
-                        
-                        variables.navigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        variables.pinnedNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        variables.favoritesNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        */
-                        
-                        variables.initialLoadDone = true
-                    }
-                    
-                    
-                    .ignoresSafeArea()
-                }
-            }.task {
-                if spaces.count <= 0 {
-                    await modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled", spaceIcon: "circle.fill", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                }
-            }
-        }
-        else {
-            NavigationStack {
-                //TabOverview(urls: ["https://apple.com", "https://google.com", "https://arc.net", "https://apple.com"])
-                TabOverview(selectedSpaceIndex: $selectedSpaceIndex)
-            }
-        }
-            #else
-            GeometryReader { geo in
-                if spaces.count > 0 {
-                    ZStack {
-                        ZStack {
+                                                    .padding(15)
+                                                    .frame(width: 300)
+                                                    .background(content: {
 #if !os(visionOS)
-                            if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
-                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                            }
-                            else {
-                                LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                            }
-#endif
-                            
-                            if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                Color.black.opacity(0.5)
-                            }
-                            
-                            
-                            HStack(spacing: 0) {
-                                //if UIDevice.current.userInterfaceIdiom != .phone {
-                                if true {
-                                    if settings.sidebarLeft {
-                                        if settings.showBorder {
-                                            ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
-                                                .disabled(true)
-                                        }
-                                        
-                                        if settings.swipingSpaces {
-                                            PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
-                                        }
-                                        else {
-                                            VStack {
-                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                    .padding([.top, .horizontal], 5)
-                                                
-                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                
-                                                HStack {
-                                                    Button {
-                                                        variables.showSettings.toggle()
-                                                    } label: {
-                                                        ZStack {
-                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                            
-                                                            Image(systemName: "gearshape")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.settingsButtonHover = true
+                                                        if settings.sidebarLeft {
+                                                            LinearGradient(colors: [variables.startColor, Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
+                                                                .opacity(1.0)
+                                                            if selectedSpaceIndex < spaces.count {
+                                                                if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
+                                                                    LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
                                                                 }
-                                                                else {
-                                                                    variables.settingsButtonHover = false
-                                                                }
-                                                            })
-                                                    }.buttonStyle(.plain)
-                                                    .sheet(isPresented: $variables.showSettings) {
-                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                .presentationSizing(.form)
+                                                            }
                                                         } else {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                            LinearGradient(colors: [Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex)), variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
+                                                                .opacity(1.0)
+                                                            if selectedSpaceIndex < spaces.count {
+                                                                if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
+                                                                    LinearGradient(colors: [Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex)), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
+                                                                }
+                                                            }
                                                         }
-                                                    }
-                                                    Spacer()
-                                                    
-                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                    
-                                                    Button(action: {
-                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                    }, label: {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            Color(.white)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
 #endif
-                                                            
-                                                            Image(systemName: "plus")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                }
-                                                                else {
-                                                                    variables.hoverSpace = ""
-                                                                }
-                                                            })
-                                                    }).buttonStyle(.plain)
-                                                }
-                                            }.frame(width: 300)
-                                        }
-                                    }
-                                }
-                                HStack {
-                                    GeometryReader { webGeo in
-                                        ZStack {
-                                            Color.white
-                                                .opacity(0.4)
-                                                .cornerRadius(10)
-                                            
-                                            
-                                            //MARK: - WebView
-                                            CurrentWebView(webGeo: webGeo)
-                                                .overlay(content: {
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        loadingIndicators(for: variables.navigationState.selectedWebView?.isLoading)
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        loadingIndicators(for: variables.pinnedNavigationState.selectedWebView?.isLoading)
-                                                    }
-                                                    else {
-                                                        loadingIndicators(for: variables.favoritesNavigationState.selectedWebView?.isLoading)
-                                                    }
-                                                })
-                                            
-                                            if !settings.swipeNavigationDisabled {
-                                                if (variables.selectedTabLocation == "favoriteTabs" && variables.favoritesNavigationState.selectedWebView != nil) ||
-                                                    (variables.selectedTabLocation == "pinnedTabs" && variables.pinnedNavigationState.selectedWebView != nil) ||
-                                                    (variables.selectedTabLocation == "tabs" && variables.navigationState.selectedWebView != nil) {
-                                                    HStack(alignment: .center, spacing: 0) {
-                                                        navigationButton(imageName: "arrow.left", action: goBack)
-                                                            .padding(.trailing, 30)
                                                         
-                                                        Spacer()
-                                                            .frame(width: webGeo.size.width)
-                                                        
-                                                        navigationButton(imageName: "arrow.right", action: goForward)
-                                                            .padding(.leading, 30)
-                                                        
-                                                    }
-                                                    .frame(width: webGeo.size.width)
-                                                    .offset(x: variables.navigationOffset)
-                                                }
-                                            }
-                                            
-                                            if variables.auraTab == "dashboard" && variables.selectedTabLocation == "" {
-                                                Dashboard(startHexSpace: spaces[selectedSpaceIndex].startHex, endHexSpace: spaces[selectedSpaceIndex].endHex)
+                                                        if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
+                                                            Color.black.opacity(0.5)
+                                                        }
+                                                    })
                                                     .cornerRadius(10)
-                                                    .clipped()
-                                            }
-                                            
-                                            Spacer()
-                                                .frame(width: 20)
-                                            
-                                            HStack {
-                                                Button {
-                                                    Task {
-                                                        await hideSidebar.toggle()
-                                                    }
-                                                    
-                                                    withAnimation {
-                                                        if !hideSidebar {
-                                                            if variables.selectedTabLocation == "tabs" {
-                                                                Task {
-                                                                    await variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "pinnedTabs" {
-                                                                Task {
-                                                                    await variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "favoriteTabs" {
-                                                                Task {
-                                                                    await variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                                }
-                                                            }
-                                                        } else {
-                                                            if variables.selectedTabLocation == "tabs" {
-                                                                Task {
-                                                                    await variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "pinnedTabs" {
-                                                                Task {
-                                                                    await variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                            else if variables.selectedTabLocation == "favoriteTabs" {
-                                                                Task {
-                                                                    await variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-340, height: geo.size.height))
-                                                                }
-                                                            }
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("s", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.goBack()
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.goBack()
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.goBack()
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("[", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.goForward()
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.goForward()
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.goForward()
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("]", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    variables.reloadRotation += 360
-                                                    
-                                                    if variables.selectedTabLocation == "tabs" {
-                                                        variables.navigationState.selectedWebView?.reload()
-                                                        variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.navigationState.selectedWebView = variables.navigationState.selectedWebView
-                                                        //navigationState.currentURL = navigationState.currentURL
-                                                        
-                                                        if let unwrappedURL = variables.navigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        variables.pinnedNavigationState.selectedWebView?.reload()
-                                                        variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.pinnedNavigationState.selectedWebView = variables.pinnedNavigationState.selectedWebView
-                                                        
-                                                        if let unwrappedURL = variables.pinnedNavigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "favoriteTabs" {
-                                                        variables.favoritesNavigationState.selectedWebView?.reload()
-                                                        variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: geo.size.width-40, height: geo.size.height))
-                                                        
-                                                        variables.favoritesNavigationState.selectedWebView = variables.favoritesNavigationState.selectedWebView
-                                                        
-                                                        if let unwrappedURL = variables.favoritesNavigationState.currentURL {
-                                                            variables.searchInSidebar = unwrappedURL.absoluteString
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("r", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    if (variables.navigationState.selectedWebView == nil) && (variables.pinnedNavigationState.selectedWebView == nil) && (variables.favoritesNavigationState.selectedWebView == nil) {
-                                                        variables.tabBarShown.toggle()
-                                                        variables.commandBarShown = false
-                                                        print("Showing Tab Bar")
-                                                    }
-                                                    else {
-                                                        if variables.selectedTabLocation == "pinnedTabs" {
-                                                            variables.searchInSidebar = unformatURL(url: variables.pinnedNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }
-                                                        else if variables.selectedTabLocation == "favoriteTabs" {
-                                                            variables.searchInSidebar = unformatURL(url: variables.favoritesNavigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }else {
-                                                            variables.searchInSidebar = unformatURL(url: variables.navigationState.selectedWebView?.url?.absoluteString ?? variables.searchInSidebar)
-                                                        }
-                                                        variables.commandBarShown.toggle()
-                                                        variables.tabBarShown = false
-                                                        print("Showing Command Bar")
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("l", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    if variables.selectedTabLocation == "favoriteTabs" {
-                                                        if let index = variables.favoritesNavigationState.webViews.firstIndex(of: variables.favoritesNavigationState.selectedWebView ?? WKWebView()) {
-                                                            favoriteRemoveTab(at: index)
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "pinnedTabs" {
-                                                        if let index = variables.pinnedNavigationState.webViews.firstIndex(of: variables.pinnedNavigationState.selectedWebView ?? WKWebView()) {
-                                                            pinnedRemoveTab(at: index)
-                                                        }
-                                                    }
-                                                    else if variables.selectedTabLocation == "tabs" {
-                                                        if let index = variables.navigationState.webViews.firstIndex(of: variables.navigationState.selectedWebView ?? WKWebView()) {
-                                                            removeTab(at: index)
-                                                        }
-                                                    }
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("w", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                                Button {
-                                                    variables.tabBarShown.toggle()
-                                                    variables.commandBarShown = false
-                                                } label: {
-                                                    
-                                                }.keyboardShortcut("t", modifiers: .command)
-                                                    .buttonStyle(.plain)
-                                                
-                                                
-                                            }
-                                        }
-                                        .gesture(
-                                            DragGesture()
-                                                .onChanged { value in
-                                                    if !settings.swipeNavigationDisabled {
-                                                        let startLocation = value.startLocation.x
-                                                        let width = webGeo.size.width
-                                                        
-                                                        if startLocation < 100 || startLocation > (width - 100) {
-                                                            let newOffset = value.translation.width
-                                                            if abs(newOffset) <= 150 {
-                                                                variables.navigationOffset = newOffset
-                                                            } else {
-                                                                variables.navigationOffset = newOffset > 0 ? 150 : -150
-                                                            }
-                                                            if abs(newOffset) > 100 {
-                                                                if !variables.arrowImpactOnce {
-                                                                    heavyHaptics()
-                                                                    
-                                                                    variables.arrowImpactOnce = true
-                                                                }
-                                                                
-                                                                withAnimation(.linear(duration: 0.3)) {
-                                                                    variables.navigationArrowColor = true
-                                                                }
-                                                            } else {
-                                                                if variables.arrowImpactOnce {
-                                                                    heavyHaptics()
-                                                                    
-                                                                    variables.arrowImpactOnce = false
-                                                                }
-                                                                withAnimation(.linear(duration: 0.3)) {
-                                                                    variables.navigationArrowColor = false
-                                                                }
-                                                            }
-                                                        }
-                                                    }
+                                                    .shadow(color: .black.opacity(0.25), radius: 10, x: 0, y: 0)
                                                 }
-                                                .onEnded { value in
-                                                    if !settings.swipeNavigationDisabled {
-                                                        variables.arrowImpactOnce = false
-                                                        let startLocation = value.startLocation.x
-                                                        let width = webGeo.size.width
-                                                        
-                                                        if startLocation < 150 || startLocation > (width - 150) {
-                                                            if variables.navigationOffset >= 100 {
-                                                                goBack()
-                                                            } else if variables.navigationOffset < -100 {
-                                                                goForward()
-                                                            }
-                                                            
-                                                            withAnimation(.linear(duration: 0.25)) {
-                                                                variables.navigationOffset = 0
-                                                                variables.navigationArrowColor = false
-                                                            }
-                                                        }
-                                                    }
-                                                }
-                                        )
-                                        .onChange(of: webGeo.size) {
-                                            variables.navigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                            
-                                            variables.pinnedNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                            
-                                            variables.favoritesNavigationState.selectedWebView?.frame = CGRect(origin: .zero, size: CGSize(width: webGeo.size.width, height: webGeo.size.height))
-                                        }
-                                    }
-                                    .cornerRadius(10)
-                                    .clipped()
-                                    .padding(settings.sidebarLeft ? .trailing: .leading, settings.showBorder ? 12: 0)
-                                    
-                                    if variables.delayedBrowseForMe {
-                                        BrowseForMe(searchText: variables.browseForMeSearch, searchResponse: "", closeSheet: $variables.isBrowseForMe)
-                                            .frame(width: variables.isBrowseForMe ? 400: 0)
-                                            .cornerRadius(10)
-                                            .clipped()
-                                    }
-                                }.onChange(of: variables.isBrowseForMe, {
-                                    DispatchQueue.main.asyncAfter(deadline: .now() + 1, execute: {
-                                        withAnimation(.linear, {
-                                            variables.delayedBrowseForMe = variables.isBrowseForMe
-                                        })
-                                    })
-                                })
-                                //if UIDevice.current.userInterfaceIdiom != .phone {
-                                if true {
-                                    if !settings.sidebarLeft {
-                                        if settings.showBorder {
-                                            ThreeDots(hoverTinySpace: $variables.hoverTinySpace, hideSidebar: $hideSidebar)
-                                                .disabled(true)
-                                        }
-                                        
-                                        if settings.swipingSpaces {
-                                            PagedSidebar(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, fullGeo: geo)
-                                        }
-                                        else {
-                                            VStack {
-                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                    .padding([.top, .horizontal], 5)
-                                                
-                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                
-                                                HStack {
-                                                    Button {
-                                                        variables.showSettings.toggle()
-                                                    } label: {
-                                                        ZStack {
-                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                            
-                                                            Image(systemName: "gearshape")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.settingsButtonHover = true
-                                                                }
-                                                                else {
-                                                                    variables.settingsButtonHover = false
-                                                                }
-                                                            })
-                                                    }.buttonStyle(.plain)
-                                                    .sheet(isPresented: $variables.showSettings) {
-                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                .presentationSizing(.form)
-                                                        } else {
-                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                        }
-                                                    }
-                                                    Spacer()
-                                                    
-                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                    
-                                                    Button(action: {
-                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                    }, label: {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            Color(.white)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
-#endif
-                                                            
-                                                            Image(systemName: "plus")
-                                                                .resizable()
-                                                                .scaledToFit()
-                                                                .frame(width: 20, height: 20)
-                                                                .foregroundStyle(variables.textColor)
-                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                            
-                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                            .hoverEffect(.lift)
-                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                            .onHover(perform: { hovering in
-                                                                if hovering {
-                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                }
-                                                                else {
-                                                                    variables.hoverSpace = ""
-                                                                }
-                                                            })
-                                                    }).buttonStyle(.plain)
-                                                }
-                                            }.frame(width: 300)
-                                        }
-                                    }
-                                }
-                            }
-                            .padding(.trailing, settings.showBorder ? 10: 0)
-                            .padding(.vertical, settings.showBorder ? 25: 0)
-                            .onAppear {
-                                if let savedStartColor = getColor(forKey: "startColorHex") {
-                                    variables.startColor = savedStartColor
-                                }
-                                if let savedEndColor = getColor(forKey: "endColorHex") {
-                                    variables.endColor = savedEndColor
-                                }
-                                if let savedTextColor = getColor(forKey: "textColorHex") {
-                                    variables.textColor = savedTextColor
-                                }
-                                
-                                variables.spaceIcons = UserDefaults.standard.dictionary(forKey: "spaceIcons") as? [String: String]
-                            }
-                            
-                            if variables.tabBarShown || variables.commandBarShown || variables.tapSidebarShown {
-                                Button(action: {
-                                    variables.tabBarShown = false
-                                    variables.commandBarShown = false
-                                    variables.tapSidebarShown = false
-                                }, label: {
-                                    Color.white.opacity(0.0001)
-                                }).buttonStyle(.plain)
-#if !os(macOS)
-                                    .hoverEffectDisabled(true)
-                                #endif
-                            }
-                            
-                            if hideSidebar {
-                                HStack {
-                                    if !settings.sidebarLeft {
-                                        Spacer()
-                                    }
-                                    
-                                    ZStack {
-                                        Color.white.opacity(0.00001)
-                                            .frame(width: 35)
-                                            .onTapGesture {
-                                                variables.tapSidebarShown = true
-                                                
-                                            }
-                                        
-                                        HStack {
-                                            //                                        PagedSidebar(selectedTabLocation: $selectedTabLocation, navigationState: navigationState, pinnedNavigationState: pinnedNavigationState, favoritesNavigationState: favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
-                                            
-                                            //if UIDevice.current.userInterfaceIdiom != .phone {
-                                            if true {
-                                                VStack {
-                                                    ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                        .padding([.top, .horizontal], 5)
-                                                    
-                                                    //Sidebar(selectedTabLocation: $selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $searchInSidebar, commandBarShown: $commandBarShown, tabBarShown: $tabBarShown, startColor: $startColor, endColor: $endColor, textColor: $textColor, hoverSpace: $hoverSpace, showSettings: $showSettings, geo: geo)
-                                                    Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                    
-                                                    HStack {
-                                                        Button {
-                                                            variables.showSettings.toggle()
-                                                        } label: {
-                                                            ZStack {
-                                                                HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                                
-                                                                Image(systemName: "gearshape")
-                                                                    .resizable()
-                                                                    .scaledToFit()
-                                                                    .frame(width: 20, height: 20)
-                                                                    .foregroundStyle(variables.textColor)
-                                                                    .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                                
-                                                            }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                                .hoverEffect(.lift)
-                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                                .onHover(perform: { hovering in
-                                                                    if hovering {
-                                                                        variables.settingsButtonHover = true
-                                                                    }
-                                                                    else {
-                                                                        variables.settingsButtonHover = false
-                                                                    }
-                                                                })
-                                                        }.buttonStyle(.plain)
-                                                        .sheet(isPresented: $variables.showSettings) {
-                                                            if #available(iOS 18.0, visionOS 2.0, *) {
-                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                    .presentationSizing(.form)
-                                                            } else {
-                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                            }
-                                                        }
-                                                        Spacer()
-                                                        
-                                                        SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                        
-                                                        Button(action: {
-                                                            modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                        }, label: {
+                                                else {
+                                                    Color.clear
+                                                        .sheet(isPresented: $variables.tapSidebarShown) {
                                                             ZStack {
 #if !os(visionOS)
-                                                                Color(.white)
-                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
+                                                                if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
+                                                                    LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
+                                                                }
+                                                                else {
+                                                                    LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
+                                                                }
 #endif
                                                                 
-                                                                Image(systemName: "plus")
-                                                                    .resizable()
-                                                                    .scaledToFit()
-                                                                    .frame(width: 20, height: 20)
-                                                                    .foregroundStyle(variables.textColor)
-                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
+                                                                if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
+                                                                    Color.black.opacity(0.5)
+                                                                }
                                                                 
-                                                            }.frame(width: 40, height: 40).cornerRadius(7)
+                                                                VStack {
+                                                                    ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
+                                                                        .padding([.top, .horizontal], 5)
+                                                                    
+                                                                    Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
+                                                                    
+                                                                    HStack {
+                                                                        Button {
+                                                                            variables.showSettings.toggle()
+                                                                        } label: {
+                                                                            ZStack {
+                                                                                HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
+                                                                                
+                                                                                Image(systemName: "gearshape")
+                                                                                    .resizable()
+                                                                                    .scaledToFit()
+                                                                                    .frame(width: 20, height: 20)
+                                                                                    .foregroundStyle(variables.textColor)
+                                                                                    .opacity(variables.settingsButtonHover ? 1.0: 0.5)
+                                                                                
+                                                                            }.frame(width: 40, height: 40).cornerRadius(7)
 #if !os(visionOS) && !os(macOS)
-                                                                .hoverEffect(.lift)
-                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+                                                                                .hoverEffect(.lift)
+                                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
 #endif
-                                                                .onHover(perform: { hovering in
-                                                                    if hovering {
-                                                                        variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                    }
-                                                                    else {
-                                                                        variables.hoverSpace = ""
-                                                                    }
-                                                                })
-                                                        }).buttonStyle(.plain)
-                                                    }
-                                                }
-                                                .padding(15)
-                                                .frame(width: 300)
-                                                .background(content: {
+                                                                                .onHover(perform: { hovering in
+                                                                                    if hovering {
+                                                                                        variables.settingsButtonHover = true
+                                                                                    }
+                                                                                    else {
+                                                                                        variables.settingsButtonHover = false
+                                                                                    }
+                                                                                })
+                                                                        }
+                                                                        .sheet(isPresented: $variables.showSettings) {
+                                                                            if #available(iOS 18.0, visionOS 2.0, *) {
+                                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                                                    .presentationSizing(.form)
+                                                                            } else {
+                                                                                NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
+                                                                            }
+                                                                        }
+                                                                        Spacer()
+                                                                        
+                                                                        SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
+                                                                        
+                                                                        Button(action: {
+                                                                            modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
+                                                                        }, label: {
+                                                                            ZStack {
 #if !os(visionOS)
-                                                    if settings.sidebarLeft {
-                                                        LinearGradient(colors: [variables.startColor, Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            .opacity(1.0)
-                                                        if selectedSpaceIndex < spaces.count {
-                                                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex))], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-                                                        }
-                                                    } else {
-                                                        LinearGradient(colors: [Color(hex: averageHexColor(hex1: variables.startHex, hex2: variables.endHex)), variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            .opacity(1.0)
-                                                        if selectedSpaceIndex < spaces.count {
-                                                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                                                LinearGradient(colors: [Color(hex: averageHexColor(hex1: spaces[selectedSpaceIndex].startHex, hex2: spaces[selectedSpaceIndex].endHex)), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-                                                        }
-                                                    }
+                                                                                Color(.white)
+                                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
 #endif
-                                                    
-                                                    if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                                        Color.black.opacity(0.5)
-                                                    }
-                                                })
-                                                .cornerRadius(10)
-                                                .shadow(color: .black.opacity(0.25), radius: 10, x: 0, y: 0)
+                                                                                
+                                                                                Image(systemName: "plus")
+                                                                                    .resizable()
+                                                                                    .scaledToFit()
+                                                                                    .frame(width: 20, height: 20)
+                                                                                    .foregroundStyle(variables.textColor)
+                                                                                    .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
+                                                                                
+                                                                            }.frame(width: 40, height: 40).cornerRadius(7)
+#if !os(visionOS) && !os(macOS)
+                                                                                .hoverEffect(.lift)
+                                                                                .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
+#endif
+                                                                                .onHover(perform: { hovering in
+                                                                                    if hovering {
+                                                                                        variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
+                                                                                    }
+                                                                                    else {
+                                                                                        variables.hoverSpace = ""
+                                                                                    }
+                                                                                })
+                                                                        })
+                                                                    }
+                                                                }
+                                                                .padding(15)
+                                                                .frame(width: 300)
+                                                                .cornerRadius(10)
+                                                            }
+                                                        }
+                                                }
+                                                
+                                                Spacer()
+                                            }.padding(40)
+                                                .padding(.leading, 30)
+                                                .frame(width: variables.hoveringSidebar || variables.tapSidebarShown ? 350: 0)
+                                                .offset(x: variables.hoveringSidebar || variables.tapSidebarShown ? 0: settings.sidebarLeft ? -350: 300)
+                                                .clipped()
+                                            
+                                        }.onHover(perform: { hovering in
+                                            if hovering {
+                                                variables.hoveringSidebar = true
                                             }
                                             else {
-                                                Color.clear
-                                                    .sheet(isPresented: $variables.tapSidebarShown) {
-                                                        ZStack {
-#if !os(visionOS)
-                                                            if selectedSpaceIndex < spaces.count && (!spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty) {
-                                                                LinearGradient(colors: [Color(hex: spaces[selectedSpaceIndex].startHex), Color(hex: spaces[selectedSpaceIndex].endHex)], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-                                                            else {
-                                                                LinearGradient(colors: [variables.startColor, variables.endColor], startPoint: .bottomLeading, endPoint: .topTrailing).ignoresSafeArea()
-                                                            }
-#endif
-                                                            
-                                                            if settings.prefferedColorScheme == "dark" || (settings.prefferedColorScheme == "automatic" && colorScheme == .dark) {
-                                                                Color.black.opacity(0.5)
-                                                            }
-                                                            
-                                                            VStack {
-                                                                ToolbarButtonsView(selectedTabLocation: $variables.selectedTabLocation, navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, geo: geo).frame(height: 40)
-                                                                    .padding([.top, .horizontal], 5)
-                                                                
-                                                                Sidebar(selectedTabLocation: $variables.selectedTabLocation, hideSidebar: $hideSidebar, searchInSidebar: $variables.searchInSidebar, commandBarShown: $variables.commandBarShown, tabBarShown: $variables.tabBarShown, startColor: $variables.startColor, endColor: $variables.endColor, textColor: $variables.textColor, hoverSpace: $variables.hoverSpace, showSettings: $variables.showSettings, geo: geo)
-                                                                
-                                                                HStack {
-                                                                    Button {
-                                                                        variables.showSettings.toggle()
-                                                                    } label: {
-                                                                        ZStack {
-                                                                            HoverButtonDisabledVision(hoverInteraction: $variables.settingsButtonHover)
-                                                                            
-                                                                            Image(systemName: "gearshape")
-                                                                                .resizable()
-                                                                                .scaledToFit()
-                                                                                .frame(width: 20, height: 20)
-                                                                                .foregroundStyle(variables.textColor)
-                                                                                .opacity(variables.settingsButtonHover ? 1.0: 0.5)
-                                                                            
-                                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                                            .hoverEffect(.lift)
-                                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                                            .onHover(perform: { hovering in
-                                                                                if hovering {
-                                                                                    variables.settingsButtonHover = true
-                                                                                }
-                                                                                else {
-                                                                                    variables.settingsButtonHover = false
-                                                                                }
-                                                                            })
-                                                                    }.buttonStyle(.plain)
-                                                                    .sheet(isPresented: $variables.showSettings) {
-                                                                        if #available(iOS 18.0, visionOS 2.0, *) {
-                                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                                .presentationSizing(.form)
-                                                                        } else {
-                                                                            NewSettings(presentSheet: $variables.showSettings, startHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].startHex: startHex, endHex: (!spaces[selectedSpaceIndex].startHex.isEmpty) ? spaces[selectedSpaceIndex].endHex: endHex)
-                                                                        }
-                                                                    }
-                                                                    Spacer()
-                                                                    
-                                                                    SpacePicker(navigationState: variables.navigationState, pinnedNavigationState: variables.pinnedNavigationState, favoritesNavigationState: variables.favoritesNavigationState, currentSpace: $currentSpace, selectedSpaceIndex: $selectedSpaceIndex)
-                                                                    
-                                                                    Button(action: {
-                                                                        modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled \(spaces.count)", spaceIcon: "scribble.variable", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
-                                                                    }, label: {
-                                                                        ZStack {
-#if !os(visionOS)
-                                                                            Color(.white)
-                                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 0.5: 0.0)
-#endif
-                                                                            
-                                                                            Image(systemName: "plus")
-                                                                                .resizable()
-                                                                                .scaledToFit()
-                                                                                .frame(width: 20, height: 20)
-                                                                                .foregroundStyle(variables.textColor)
-                                                                                .opacity(variables.hoverSpace == "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable" ? 1.0: 0.5)
-                                                                            
-                                                                        }.frame(width: 40, height: 40).cornerRadius(7)
-#if !os(visionOS) && !os(macOS)
-                                                                            .hoverEffect(.lift)
-                                                                            .hoverEffectDisabled(!settings.hoverEffectsAbsorbCursor)
-#endif
-                                                                            .onHover(perform: { hovering in
-                                                                                if hovering {
-                                                                                    variables.hoverSpace = "veryLongTextForHoveringOnPlusSignSoIDontHaveToUseAnotherVariable"
-                                                                                }
-                                                                                else {
-                                                                                    variables.hoverSpace = ""
-                                                                                }
-                                                                            })
-                                                                    }).buttonStyle(.plain)
-                                                                }
-                                                            }
-                                                            .padding(15)
-                                                            .frame(width: 300)
-                                                            .cornerRadius(10)
-                                                        }
-                                                    }
+                                                variables.hoveringSidebar = false
                                             }
-                                            
-                                            Spacer()
-                                        }.padding(40)
-                                            .padding(.leading, 30)
-                                            .frame(width: variables.hoveringSidebar || variables.tapSidebarShown ? 350: 0)
-                                            .offset(x: variables.hoveringSidebar || variables.tapSidebarShown ? 0: settings.sidebarLeft ? -350: 300)
-                                            .clipped()
+                                        })
                                         
-                                    }.onHover(perform: { hovering in
-                                        if hovering {
-                                            variables.hoveringSidebar = true
+                                        if settings.sidebarLeft {
+                                            Spacer()
+                                        }
+                                    }.animation(.default)
+                                }
+                            }
+                            .onOpenURL { url in
+                                if url.absoluteString.starts(with: "aura://") {
+                                    //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: "https\(url.absoluteString.dropFirst(4))")!))
+                                }
+                                else {
+                                    //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: url.absoluteString)!))
+                                    
+                                    spaces[selectedSpaceIndex].tabUrls.append(url.absoluteString)
+                                }
+                                print("Url:")
+                                print(url)
+                            }
+                            //MARK: - Tabbar
+                            if variables.tabBarShown {
+                                CommandBar(commandBarText: $variables.newTabSearch, searchSubmitted: $variables.commandBarSearchSubmitted, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
+                                    .onChange(of: variables.commandBarSearchSubmitted) { thing in
+                                        
+                                        print("Search submitted")
+                                        
+                                        variables.browseForMeSearch = variables.newTabSearch
+                                        
+                                        if !variables.newTabSearch.starts(with: "aura://") {
+                                            variables.auraTab = ""
+                                            //variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: formatURL(from: variables.newTabSearch))!))
+                                            
+                                            print(variables.newTabSearch)
+                                            
+                                            var temporaryUrls = spaces[selectedSpaceIndex].tabUrls
+                                            
+                                            print("temporaryUrls:")
+                                            print(temporaryUrls)
+                                            
+                                            var formattedUrl = formatURL(from: variables.newTabSearch)
+                                            
+                                            print("formattedUrl:")
+                                            print(formattedUrl)
+                                            
+                                            temporaryUrls.append(formattedUrl)
+                                            
+                                            print("temporaryUrls - changed:")
+                                            print(temporaryUrls)
+                                            spaces[selectedSpaceIndex].tabUrls = temporaryUrls
+                                            
+                                            print("tabUrls:")
+                                            print(spaces[selectedSpaceIndex].tabUrls)
+                                            
+                                            do {
+                                                try modelContext.save()
+                                            } catch {
+                                                print(error.localizedDescription)
+                                            }
+                                            print(spaces[selectedSpaceIndex].tabUrls)
                                         }
                                         else {
-                                            variables.hoveringSidebar = false
-                                        }
-                                    })
-                                    
-                                    if settings.sidebarLeft {
-                                        Spacer()
-                                    }
-                                }.animation(.default)
-                            }
-                        }
-                        .onOpenURL { url in
-                            if url.absoluteString.starts(with: "aura://") {
-                                variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: "https\(url.absoluteString.dropFirst(4))")!))
-                            }
-                            else {
-                                variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: url.absoluteString)!))
-                            }
-                            print("Url:")
-                            print(url)
-                        }
-                        //MARK: - Tabbar
-                        if variables.tabBarShown {
-                            CommandBar(commandBarText: $variables.newTabSearch, searchSubmitted: $variables.commandBarSearchSubmitted, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
-                                .onChange(of: variables.commandBarSearchSubmitted) { thing in
-                                    
-                                    variables.browseForMeSearch = variables.newTabSearch
-                                    
-                                    if !variables.newTabSearch.starts(with: "aura://") {
-                                        variables.auraTab = ""
-                                        variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: formatURL(from: variables.newTabSearch))!))
-                                    }
-                                    else {
-                                        if variables.newTabSearch.contains("dashboard") {
-                                            variables.navigationState.selectedWebView = nil
-                                            variables.pinnedNavigationState.selectedWebView = nil
-                                            variables.favoritesNavigationState.selectedWebView = nil
-                                            
-                                            variables.auraTab = "dashboard"
-                                            variables.selectedTabLocation = ""
-                                        }
-                                        if variables.newTabSearch.contains("settings") {
-                                            variables.showSettings = true
-                                        }
-                                    }
-                                    
-                                    variables.tabBarShown = false
-                                    variables.commandBarSearchSubmitted = false
-                                    variables.newTabSearch = ""
-                                    
-                                    print("Saving Tabs")
-                                    
-                                    saveSpaceData()
-                                }
-                            //}
-                        }
-                        
-                        //MARK: - Command Bar
-                        else if variables.commandBarShown {
-                            CommandBar(commandBarText: $variables.searchInSidebar, searchSubmitted: $variables.commandBarSearchSubmitted2, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
-                                .onChange(of: variables.navigationState.currentURL, {
-                                    if let unwrappedURL = variables.navigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 1")
-                                })
-                                .onChange(of: variables.pinnedNavigationState.currentURL, {
-                                    if let unwrappedURL = variables.pinnedNavigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 2")
-                                })
-                                .onChange(of: variables.favoritesNavigationState.currentURL, {
-                                    if let unwrappedURL = variables.favoritesNavigationState.currentURL {
-                                        variables.searchInSidebar = unwrappedURL.absoluteString
-                                    }
-                                    print("Changing searchInSidebar - 3")
-                                })
-                                .onChange(of: variables.commandBarSearchSubmitted2) { thing in
-                                    
-                                    //variables.navigationState.currentURL = URL(string: formatURL(from: newTabSearch))!
-                                    //variables.navigationState.selectedWebView?.load(URLRequest(url: URL(formatURL(from: newTabSearch))!))
-                                    Task {
-                                        await variables.searchInSidebar = formatURL(from: variables.searchInSidebar)
-                                        if let url = URL(string: variables.searchInSidebar) {
-                                            // Create a URLRequest object
-                                            let request = URLRequest(url: url)
-                                            
-                                            if variables.selectedTabLocation == "tabs" {
-                                                await variables.navigationState.selectedWebView?.load(request)
+                                            if variables.newTabSearch.contains("dashboard") {
+                                                variables.navigationState.selectedWebView = nil
+                                                variables.pinnedNavigationState.selectedWebView = nil
+                                                variables.favoritesNavigationState.selectedWebView = nil
+                                                
+                                                variables.auraTab = "dashboard"
+                                                variables.selectedTabLocation = ""
                                             }
-                                            if variables.selectedTabLocation == "pinnedTabs" {
-                                                await variables.pinnedNavigationState.selectedWebView?.load(request)
+                                            if variables.newTabSearch.contains("settings") {
+                                                variables.showSettings = true
                                             }
-                                            if variables.selectedTabLocation == "favoriteTabs" {
-                                                await variables.favoritesNavigationState.selectedWebView?.load(request)
-                                            }
-                                            
-                                            print("Updated URL String")
-                                        } else {
-                                            print("Invalid URL string")
                                         }
                                         
-                                        saveSpaceData()
+                                        variables.tabBarShown = false
+                                        variables.commandBarSearchSubmitted = false
+                                        variables.newTabSearch = ""
+                                        
+                                        print("Saving Tabs")
+                                        
+                                        //saveSpaceData()
                                     }
-                                    
-                                    
-                                    variables.commandBarShown = false
-                                    variables.tabBarShown = false
-                                    variables.commandBarSearchSubmitted2 = false
-                                    variables.newTabSearch = ""
+                                //}
+                            }
+                            
+                            //MARK: - Command Bar
+                            else if variables.commandBarShown {
+                                CommandBar(commandBarText: $variables.searchInSidebar, searchSubmitted: $variables.commandBarSearchSubmitted2, collapseHeightAnimation: $variables.commandBarCollapseHeightAnimation, isBrowseForMe: $variables.isBrowseForMe)
+                                    .onChange(of: variables.navigationState.currentURL, {
+                                        if let unwrappedURL = variables.navigationState.currentURL {
+                                            variables.searchInSidebar = unwrappedURL.absoluteString
+                                        }
+                                        print("Changing searchInSidebar - 1")
+                                    })
+                                    .onChange(of: variables.pinnedNavigationState.currentURL, {
+                                        if let unwrappedURL = variables.pinnedNavigationState.currentURL {
+                                            variables.searchInSidebar = unwrappedURL.absoluteString
+                                        }
+                                        print("Changing searchInSidebar - 2")
+                                    })
+                                    .onChange(of: variables.favoritesNavigationState.currentURL, {
+                                        if let unwrappedURL = variables.favoritesNavigationState.currentURL {
+                                            variables.searchInSidebar = unwrappedURL.absoluteString
+                                        }
+                                        print("Changing searchInSidebar - 3")
+                                    })
+                                    .onChange(of: variables.commandBarSearchSubmitted2) { thing in
+                                        
+                                        //variables.navigationState.currentURL = URL(string: formatURL(from: newTabSearch))!
+                                        //variables.navigationState.selectedWebView?.load(URLRequest(url: URL(formatURL(from: newTabSearch))!))
+                                        Task {
+                                            await variables.searchInSidebar = formatURL(from: variables.searchInSidebar)
+                                            if let url = URL(string: variables.searchInSidebar) {
+                                                // Create a URLRequest object
+                                                let request = URLRequest(url: url)
+                                                
+                                                if variables.selectedTabLocation == "tabs" {
+                                                    await variables.navigationState.selectedWebView?.load(request)
+                                                }
+                                                if variables.selectedTabLocation == "pinnedTabs" {
+                                                    await variables.pinnedNavigationState.selectedWebView?.load(request)
+                                                }
+                                                if variables.selectedTabLocation == "favoriteTabs" {
+                                                    await variables.favoritesNavigationState.selectedWebView?.load(request)
+                                                }
+                                                
+                                                print("Updated URL String")
+                                            } else {
+                                                print("Invalid URL string")
+                                            }
+                                        }
+                                        
+                                        
+                                        variables.commandBarShown = false
+                                        variables.tabBarShown = false
+                                        variables.commandBarSearchSubmitted2 = false
+                                        variables.newTabSearch = ""
+                                    }
+                            }
+                            
+                            
+                            if launchingAnimation && settings.launchAnimation {
+                                Launch_Animation()
+                            }
+                        }
+                        /*.sheet(isPresented: $isBrowseForMe, content: {
+                         VStack {
+                         if UIDevice.current.userInterfaceIdiom != .phone {
+                         BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
+                         .frame(width: geo.size.width * 0.7, height: geo.size.height * 0.9)
+                         .onDisappear() {
+                         isBrowseForMe = false
+                         newTabSearch = ""
+                         commandBarShown = false
+                         tabBarShown = false
+                         commandBarSearchSubmitted = false
+                         commandBarSearchSubmitted2 = false
+                         }
+                         }
+                         else {
+                         BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
+                         .onDisappear() {
+                         isBrowseForMe = false
+                         newTabSearch = ""
+                         commandBarShown = false
+                         tabBarShown = false
+                         commandBarSearchSubmitted = false
+                         commandBarSearchSubmitted2 = false
+                         }
+                         }
+                         }
+                         .interactiveDismissDisabled(true)
+                         })*/
+                        .onAppear() {
+                            DispatchQueue.main.asyncAfter(deadline: .now() + 4.6, execute: {
+                                
+                            })
+                        }
+                        .onChange(of: selectedSpaceIndex, {
+                            if variables.initialLoadDone {
+                                variables.navigationState.webViews.removeAll()
+                                
+                                var reloadAuraTabs = variables.auraTab
+                                variables.auraTab = ""
+                                DispatchQueue.main.asyncAfter(deadline: .now() + 0.00001) {
+                                    variables.auraTab = reloadAuraTabs
                                 }
-                        }
-                        
-                        
-                        if launchingAnimation && settings.launchAnimation {
-                            Launch_Animation()
-                        }
-                    }
-                    /*.sheet(isPresented: $isBrowseForMe, content: {
-                     VStack {
-                     if UIDevice.current.userInterfaceIdiom != .phone {
-                     BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
-                     .frame(width: geo.size.width * 0.7, height: geo.size.height * 0.9)
-                     .onDisappear() {
-                     isBrowseForMe = false
-                     newTabSearch = ""
-                     commandBarShown = false
-                     tabBarShown = false
-                     commandBarSearchSubmitted = false
-                     commandBarSearchSubmitted2 = false
-                     }
-                     }
-                     else {
-                     BrowseForMe(searchText: newTabSearch, searchResponse: "", closeSheet: $isBrowseForMe)
-                     .onDisappear() {
-                     isBrowseForMe = false
-                     newTabSearch = ""
-                     commandBarShown = false
-                     tabBarShown = false
-                     commandBarSearchSubmitted = false
-                     commandBarSearchSubmitted2 = false
-                     }
-                     }
-                     }
-                     .interactiveDismissDisabled(true)
-                     })*/
-                    .onAppear() {
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 4.6, execute: {
-                            
+                                
+                                if selectedSpaceIndex < spaces.count {
+                                    if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
+                                        startHex = spaces[selectedSpaceIndex].startHex
+                                        endHex = spaces[selectedSpaceIndex].startHex
+                                        
+                                        variables.startColor = Color(hex: spaces[selectedSpaceIndex].startHex)
+                                        variables.endColor = Color(hex: spaces[selectedSpaceIndex].endHex)
+                                    }
+                                }
+                                
+                                UserDefaults.standard.setValue(selectedSpaceIndex, forKey: "savedSelectedSpaceIndex")
+                            }
                         })
-                    }
-                    .onChange(of: selectedSpaceIndex, {
-                        if variables.initialLoadDone {
-                            variables.navigationState.webViews.removeAll()
-                            
-                            var reloadAuraTabs = variables.auraTab
-                            variables.auraTab = ""
-                            DispatchQueue.main.asyncAfter(deadline: .now() + 0.00001) {
-                                variables.auraTab = reloadAuraTabs
+                        .onChange(of: variables.navigationState.webViews, {
+                            saveSpaceData()
+                            print("Saving navigationState")
+                        })
+                        .onChange(of: variables.pinnedNavigationState.webViews, {
+                            saveSpaceData()
+                            print("Saving pinnedNavigationState")
+                        })
+                        .onChange(of: variables.favoritesNavigationState.webViews, {
+                            saveSpaceData()
+                            print("Saving favoritesNavigationState")
+                        })
+                        .onAppear() {
+                            if UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex") > spaces.count - 1 {
+                                selectedSpaceIndex = 0
+                            }
+                            else {
+                                selectedSpaceIndex = UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex")
                             }
                             
                             if selectedSpaceIndex < spaces.count {
@@ -2077,109 +1031,68 @@ struct ContentView: View {
                                 }
                             }
                             
-                            UserDefaults.standard.setValue(selectedSpaceIndex, forKey: "savedSelectedSpaceIndex")
-                        }
-                    })
-                    .onChange(of: variables.navigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving navigationState")
-                    })
-                    .onChange(of: variables.pinnedNavigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving pinnedNavigationState")
-                    })
-                    .onChange(of: variables.favoritesNavigationState.webViews, {
-                        saveSpaceData()
-                        print("Saving favoritesNavigationState")
-                    })
-                    .onAppear() {
-                        if UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex") > spaces.count - 1 {
-                            selectedSpaceIndex = 0
-                        }
-                        else {
-                            selectedSpaceIndex = UserDefaults.standard.integer(forKey: "savedSelectedSpaceIndex")
-                        }
-                        
-                        if selectedSpaceIndex < spaces.count {
-                            if !spaces[selectedSpaceIndex].startHex.isEmpty && !spaces[selectedSpaceIndex].endHex.isEmpty {
-                                startHex = spaces[selectedSpaceIndex].startHex
-                                endHex = spaces[selectedSpaceIndex].startHex
-                                
-                                variables.startColor = Color(hex: spaces[selectedSpaceIndex].startHex)
-                                variables.endColor = Color(hex: spaces[selectedSpaceIndex].endHex)
-                            }
+                            /*for space in spaces {
+                             if space.spaceName == currentSpace {
+                             for tab in space.tabUrls {
+                             variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://figma.com")!))
+                             }
+                             for tab in space.pinnedUrls {
+                             variables.pinnedNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://thebrowser.company")!))
+                             }
+                             for tab in space.favoritesUrls {
+                             variables.favoritesNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://arc.net")!))
+                             }
+                             }
+                             }
+                             variables.navigationState.selectedWebView = nil
+                             variables.pinnedNavigationState.selectedWebView = nil
+                             variables.favoritesNavigationState.selectedWebView = nil
+                             
+                             variables.navigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
+                             variables.pinnedNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
+                             variables.favoritesNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
+                             */
+                            
+                            variables.initialLoadDone = true
                         }
                         
-                        for space in spaces {
-                            if space.spaceName == currentSpace {
-                                for tab in space.tabUrls {
-                                    variables.navigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://figma.com")!))
-                                }
-                                for tab in space.pinnedUrls {
-                                    variables.pinnedNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://thebrowser.company")!))
-                                }
-                                for tab in space.favoritesUrls {
-                                    variables.favoritesNavigationState.createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://arc.net")!))
-                                }
-                            }
-                        }
-                        variables.navigationState.selectedWebView = nil
-                        variables.pinnedNavigationState.selectedWebView = nil
-                        variables.favoritesNavigationState.selectedWebView = nil
                         
-                        variables.navigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        variables.pinnedNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        variables.favoritesNavigationStateArray = Array(repeating: NavigationState(), count: spaces.count)
-                        
-                        //                    for spaceIndex in 0..<spaces.count {
-                        //                        for tab in spaces[spaceIndex].tabUrls {
-                        //                            variables.navigationStateArray[spaceIndex].createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://figma.com")!))
-                        //                        }
-                        //                        for tab in spaces[spaceIndex].pinnedUrls {
-                        //                            variables.pinnedNavigationStateArray[spaceIndex].createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://thebrowser.company")!))
-                        //                        }
-                        //                        for tab in spaces[spaceIndex].favoritesUrls {
-                        //                            variables.favoritesNavigationStateArray[spaceIndex].createNewWebView(withRequest: URLRequest(url: URL(string: tab) ?? URL(string: "https://arc.net")!))
-                        //                        }
-                        //                        variables.navigationStateArray[spaceIndex].selectedWebView = nil
-                        //                        variables.pinnedNavigationStateArray[spaceIndex].selectedWebView = nil
-                        //                        variables.favoritesNavigationStateArray[spaceIndex].selectedWebView = nil
-                        //                    }
-                        
-                        variables.initialLoadDone = true
+                        .ignoresSafeArea()
                     }
-                    
-                    
-                    .ignoresSafeArea()
-                }
-            }.task {
-                if spaces.count <= 0 {
-                    await modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled", spaceIcon: "circle.fill", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
+                }.task {
+                    if spaces.count <= 0 {
+                        await modelContext.insert(SpaceStorage(spaceIndex: spaces.count, spaceName: "Untitled", spaceIcon: "circle.fill", favoritesUrls: [], pinnedUrls: [], tabUrls: []))
+                    }
                 }
             }
+            else {
+                NavigationStack {
+                    //TabOverview(urls: ["https://apple.com", "https://google.com", "https://arc.net", "https://apple.com"])
+                    TabOverview(selectedSpaceIndex: $selectedSpaceIndex)
+                }
+            }
+
             
-            #endif
-            
-    }.environmentObject(variables)
+        }.environmentObject(variables)
             .environmentObject(manager)
         
     }
     
     func saveSpaceData() {
         /*let savingTodayTabs = variables.navigationState.webViews.compactMap { $0.url?.absoluteString }
-        let savingPinnedTabs = variables.pinnedNavigationState.webViews.compactMap { $0.url?.absoluteString }
-        let savingFavoriteTabs = variables.favoritesNavigationState.webViews.compactMap { $0.url?.absoluteString }
-        
-        if !spaces.isEmpty {
-            print("Saving Today Tabs: \(savingTodayTabs)")
-            spaces[selectedSpaceIndex].tabUrls = savingTodayTabs
-            print(spaces[selectedSpaceIndex].tabUrls)
-            spaces[selectedSpaceIndex].pinnedUrls = savingPinnedTabs
-            spaces[selectedSpaceIndex].favoritesUrls = savingFavoriteTabs
-        }
-        else {
-            modelContext.insert(SpaceStorage(spaceIndex: 0, spaceName: "Untitled", spaceIcon: "circle.fill", favoritesUrls: [], pinnedUrls: [], tabUrls: savingTodayTabs))
-        }*/
+         let savingPinnedTabs = variables.pinnedNavigationState.webViews.compactMap { $0.url?.absoluteString }
+         let savingFavoriteTabs = variables.favoritesNavigationState.webViews.compactMap { $0.url?.absoluteString }
+         
+         if !spaces.isEmpty {
+         print("Saving Today Tabs: \(savingTodayTabs)")
+         spaces[selectedSpaceIndex].tabUrls = savingTodayTabs
+         print(spaces[selectedSpaceIndex].tabUrls)
+         spaces[selectedSpaceIndex].pinnedUrls = savingPinnedTabs
+         spaces[selectedSpaceIndex].favoritesUrls = savingFavoriteTabs
+         }
+         else {
+         modelContext.insert(SpaceStorage(spaceIndex: 0, spaceName: "Untitled", spaceIcon: "circle.fill", favoritesUrls: [], pinnedUrls: [], tabUrls: savingTodayTabs))
+         }*/
         
         Task {
             do {
@@ -2229,9 +1142,9 @@ struct ContentView: View {
                 .foregroundStyle(Color.white)
             
         }.frame(width: 50, height: 50)
-        .gesture(TapGesture().onEnded(action))
+            .gesture(TapGesture().onEnded(action))
     }
-
+    
     private func handleDragChange(_ value: DragGesture.Value) {
         let newOffset = value.translation.width
         if abs(newOffset) <= 150 {
@@ -2249,7 +1162,7 @@ struct ContentView: View {
             }
         }
     }
-
+    
     private func handleDragEnd() {
         if variables.navigationOffset >= 100 {
             goBack()
@@ -2262,7 +1175,7 @@ struct ContentView: View {
             variables.navigationArrowColor = false
         }
     }
-
+    
     private func handleRotation() {
         if variables.offset == 0.5 {
             variables.offset = 0.0
@@ -2275,11 +1188,11 @@ struct ContentView: View {
             }
         }
     }
-
+    
     private func goBack() {
         manager.selectedWebView?.webView.goBack()
     }
-
+    
     private func goForward() {
         manager.selectedWebView?.webView.goForward()
     }
